scraper_server/requirements.txt
aiohttp
annotated-types
anyio
asyncio
azure-storage-blob
certifi
charset-normalizer
click
exceptiongroup
fastapi
h11
idna
pydantic
pydantic_core
requests
sniffio
starlette
typing_extensions
urllib3
uvicorn
